add get by Id

+++ Order/src/main/java/com/marco/order/domain/model/orderaggregate/IOrderRepository.java
@@ -7,5 +7,5 @@ public interface IOrderRepository extends IRepository<Order> {
 
     void update(Order order);
 
-    Order getById(int orderId);
+    Order getById(long orderId);
 }
